+++ app/PersonioIntegration/Widgets/Description.php
@@ -10,7 +10,12 @@ namespace PersonioIntegrationLight\PersonioIntegration\Widgets;
 // prevent direct access.
 defined( 'ABSPATH' ) || exit;
 
+use PersonioIntegrationLight\Helper;
+use PersonioIntegrationLight\PersonioIntegration\Position;
+use PersonioIntegrationLight\PersonioIntegration\PostTypes\PersonioPosition;
 use PersonioIntegrationLight\PersonioIntegration\Widget_Base;
+use PersonioIntegrationLight\Plugin\Languages;
+use PersonioIntegrationLight\Plugin\Templates;
 
 /**
  * Object to handle the description widget.
@@ -97,6 +102,52 @@ class Description extends Widget_Base {
 	 * @return string
 	 */
 	public function render( array $attributes ): string {
-		return Archive::get_instance()->render( $attributes );
+		$position = $this->get_position_by_request();
+		if ( ( $position instanceof Position && ! $position->is_valid() ) || ! ( $position instanceof Position ) ) {
+			return '';
+		}
+
+		// set actual language.
+		$position->set_lang( Languages::get_instance()->get_current_lang() );
+
+		// collect the attributes.
+		$attributes = array_merge( $attributes, array(
+			'personioid'              => absint( $position->get_personio_id() ),
+			'jobdescription_template' => empty( $attributes['template'] ) ? get_option( 'personioIntegrationTemplateJobDescription' ) : $attributes['template'],
+			'templates'               => array( 'content' ),
+		));
+
+		// generate styling.
+		if( ! empty( $attributes['styles'] ) ) {
+			Helper::add_inline_style( $attributes['styles'] );
+		}
+
+		// return the output of the template.
+		return Templates::get_instance()->get_direct_content_template( $position, $attributes );
+	}
+
+	/**
+	 * Return the list of params this widget requires.
+	 *
+	 * @return array<string,array<string,mixed>>
+	 */
+	public function get_params(): array {
+		// get the possible field values.
+		$templates = array();
+		foreach ( PersonioPosition::get_instance()->get_jobdescription_templates_via_rest_api() as $template ) {
+			$templates[] = $template['value'];
+		}
+
+		// generate the template-list.
+		$template_list = ' <code data-copied-label="' . esc_attr__( 'copied', 'wp-personio-integration' ) . '" title="' . esc_attr__( 'Click to copy this code in your clipboard', 'wp-personio-integration' ) . '">' . implode( '</code>, <code>', $templates ) . '</code>';
+
+		// return the list of params for this widget.
+		return array(
+			'template' => array(
+				'label'         => __( 'Name of chosen template, one of these values:', 'wp-personio-integration' ) . $template_list,
+				'example_value' => $templates[0],
+				'required'      => false,
+			),
+		);
 	}
 }